Move specialized test cases into own package.
